Silent SOS/requirements.txt
# Required packages for gesture detection project
opencv-python
numpy
PyYAML